ams: take three tries to edit a structure

+++ libraries/libvapours/include/vapours/ams/ams_fatal_error_context.hpp
@@ -48,7 +48,7 @@ namespace ams::impl {
         u64 module_base;
         u32 pstate;
         u32 afsr0;
-        u64 afsr1;
+        u32 afsr1;
         u32 esr;
         u64 far;
         u64 report_identifier; /* Normally just system tick. */
